+++ snapcraft/internal/repo/errors.py
@@ -14,14 +14,11 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from typing import Sequence
+from typing import List, Optional, Sequence
 
+from snapcraft.internal import errors
 from snapcraft.internal.errors import SnapcraftException
 from snapcraft.internal.os_release import OsRelease
-from ._platform import _is_deb_based
-from snapcraft.internal import errors
-
-from typing import List, Optional
 
 
 class RepoError(errors.SnapcraftError):
@@ -101,8 +98,7 @@ class PackageNotFoundError(RepoError):
     def message(self):
         message = "The package {!r} was not found.".format(self.package_name)
         # If the package was multiarch, try to help.
-        distro = OsRelease().id()
-        if _is_deb_based(distro) and ":" in self.package_name:
+        if ":" in self.package_name:
             (name, arch) = self.package_name.split(":", 2)
             if arch:
                 message += (
