+++ tests/unit/commands/__init__.py
@@ -14,17 +14,17 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import fixtures
 import json
 import subprocess
 from textwrap import dedent
 from unittest import mock
 
+import fixtures
 from click.testing import CliRunner
 
 from snapcraft import storeapi
-from snapcraft.storeapi.v2.channel_map import ChannelMap
 from snapcraft.cli._runner import run
+from snapcraft.storeapi.v2.channel_map import ChannelMap
 from tests import fixture_setup, unit
 
 _sample_keys = [
@@ -312,6 +312,6 @@ class FakeStoreCommandsBaseTestCase(CommandBaseTestCase):
 
         # Pretend that the snap command is available
         self.fake_package_installed = fixtures.MockPatch(
-            "snapcraft.internal.repo.Repo.is_package_installed", return_value=True
+            "snapcraft.internal.repo.AptRepo.is_package_installed", return_value=True
         )
         self.useFixture(self.fake_package_installed)
