+++ tests/unit/commands/test_refresh.py
@@ -20,6 +20,7 @@ from unittest import mock
 from testtools.matchers import Equals
 
 from tests.unit import TestWithFakeRemoteParts
+
 from . import CommandBaseTestCase
 
 
@@ -50,7 +51,7 @@ class RefreshCommandBaseTestCase(CommandBaseTestCase, TestWithFakeRemoteParts):
 
 
 class RefreshCommandTestCase(RefreshCommandBaseTestCase):
-    @mock.patch("snapcraft.cli.containers.repo.Repo.refresh_build_packages")
+    @mock.patch("snapcraft.cli.containers.repo.AptRepo.refresh_build_packages")
     def test_refresh(self, mock_repo_refresh):
         self.make_snapcraft_yaml()
