+++ tests/unit/plugins/v1/test_dump.py
@@ -169,7 +169,7 @@ class DumpPluginTestCase(unit.TestCase):
 
         # Even though this symlink is absolute, since it's to libc the copy
         # plugin shouldn't try to follow it or modify it.
-        libc_libs = snapcraft.repo.Repo.get_package_libraries("libc6")
+        libc_libs = snapcraft.repo.AptRepo.get_package_libraries("libc6")
 
         # We don't care which lib we're testing with, as long as it's a .so.
         libc_library_path = [lib for lib in libc_libs if ".so" in lib][0]
